Add unique item id for Bot
Fix statusColor rendering on top of the player
